refactor: add context file
